getProductDetailBySomeParameters
first version
